make "boot from USB" work with u-boot on RK3188
XXX THIS IS VERY DIRTY HACK! XXX

this only works with latest u-boot for RK3188 from
 https://github.com/linux-rockchip/u-boot-rockchip

e.g.)
 rkflashtool M 0x62000000 0x1000000 < ramdisk.img
 rkflashtool M 0x63000000 0x1000000 < kernel.img
 rkflashtool M 0x64000000 0x1000000 < parameter.img
 rkflashtool B 0x63000000 0x64000000

protocol may be incompatible with Rockchip. it may be changed in future.

you may omit rkcrc header from ramdisk and/or parameter.

if you omit rkcrc header from kernel, it will not be relocated to the
load address described in KERNEL_IMG: line in parameter. (i.e. kernel
needs to be coded to start at 0x63000000 in above case)